[DAD-389] refactor: user Service 폴더 분리
